requirements.txt
Django==4.0.4
django-admin-interface==0.19.0
django-archive==0.2.0
django-colorfield==0.6.3
django-filter==21.1
django-import-export==2.8.0
django-tinymce==3.4.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.18.1
Markdown==3.3.6
Pillow==9.1.0
PyJWT==2.3.0
python-decouple==3.6
django-crispy-forms
django-extensions
django-registration-redux
